# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "bq — Birthday Queue",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/suggestions": {
      "post": {
        "tags": [
          "convidado"
        ],
        "summary": "Suggest",
        "description": "A ORDEM DE VALIDAÇÃO É NORMATIVA (05 §3), porque decide qual mensagem a pessoa vê.\n\nE o cooldown é verificado no passo 2 mas gravado no passo 7: assim uma tentativa recusada\nnão gasta a vez (RF-09) — quem escolheu uma música de 9 minutos e levou `TOO_LONG` escolhe\noutra imediatamente, em vez de esperar 2 minutos por um erro.",
        "operationId": "suggest_api_suggestions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SuggestIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuggestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/search": {
      "get": {
        "tags": [
          "busca"
        ],
        "summary": "Search",
        "operationId": "search_api_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/karaoke/search": {
      "get": {
        "tags": [
          "karaokê"
        ],
        "summary": "Search",
        "operationId": "search_api_karaoke_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KaraokeSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/karaoke/start": {
      "post": {
        "tags": [
          "karaokê"
        ],
        "summary": "Start",
        "description": "A pessoa tocou INICIAR no próprio celular.\n\nA decisão inteira mora no maestro, sob o lock: checar aqui seria checar contra um estado que\npode mudar entre a leitura e o `_open`. Esta função só traduz a recusa para o contrato.",
        "operationId": "start_api_karaoke_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KaraokeStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KaraokeStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tv/claim": {
      "post": {
        "tags": [
          "tv"
        ],
        "summary": "Tv Claim",
        "description": "Bate a cada 10 s e diz se esta aba é dona do áudio. Ver `PartyRuntime.tv_claim`.\n\nÉ também a única evidência de que existe uma `/tv` aberta ANTES do primeiro karaokê — a\ntelemetria do vídeo só existe durante uma música, e o host precisa saber antes de a sala ficar\nolhando para uma tela preta.",
        "operationId": "tv_claim_api_tv_claim_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TvClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tv/release": {
      "post": {
        "tags": [
          "tv"
        ],
        "summary": "Tv Release",
        "description": "A /tv está fechando. Chegada por `navigator.sendBeacon`, que não espera resposta.\n\n`owner` no corpo devolvido significa \"você ainda era o dono quando isto chegou\" — ninguém lê,\ne ele existe para reusar `TvClaimOut` em vez de inventar um envelope de uma linha.",
        "operationId": "tv_release_api_tv_release_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TvClaimIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvClaimOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tv/report": {
      "post": {
        "tags": [
          "tv"
        ],
        "summary": "Tv Report",
        "description": "1 Hz enquanto canta, mais um imediato a cada mudança de estado do player.\n\n`ended` e `error` são AFIRMAÇÕES e fecham a vez na hora; `playing`/`paused` só reancoram. O\nsilêncio — a ausência de relatório — entra por outra porta (o teto do maestro) e nunca vira\n\"acabou\".",
        "operationId": "tv_report_api_tv_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TvReportIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TvReportOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/host/karaoke/start": {
      "post": {
        "tags": [
          "host"
        ],
        "summary": "Host Karaoke Start",
        "description": "O host começa a vez pela pessoa: o celular dela morreu, ou ela já está de pé na frente da\nTV com o microfone. Espelha o par `DELETE /api/suggestions/{id}` × o do host.",
        "operationId": "host_karaoke_start_api_host_karaoke_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KaraokeStartIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KaraokeStartOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "KaraokeResult": {
        "properties": {
          "videoId": {
            "type": "string",
            "title": "Videoid"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "thumbUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thumburl"
          },
          "durationMs": {
            "type": "integer",
            "title": "Durationms"
          },
          "queueable": {
            "type": "boolean",
            "title": "Queueable",
            "default": true
          },
          "blockedReason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ALREADY_QUEUED",
                  "TOO_LONG"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockedreason"
          },
          "blockedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockedby"
          }
        },
        "type": "object",
        "required": [
          "videoId",
          "title",
          "channel",
          "thumbUrl",
          "durationMs"
        ],
        "title": "KaraokeResult",
        "description": "Espelha `SearchResult`: `queueable` calculado NO SERVIDOR contra a fila de agora.\n\nAqui a decisão de produto pesa mais que na busca normal. Um karaokê recusado só na hora de\ntocar significa o nome da pessoa já no telão, o microfone na mão, e a tela dizendo que o\nvídeo não pode ser embutido — na frente de todos. Esmaecido e explicado no celular dela move\na falha para onde ela custa uma escolha, e não uma vergonha.\n\n🔴 Não tem `PLAYED_RECENTLY`: cantar \"Evidências\" e ouvir \"Evidências\" são `track_id`\ndiferentes e experiências diferentes. A janela de repetição não deve ligar as duas."
      },
      "KaraokeSearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/KaraokeResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "KaraokeSearchResponse"
      },
      "KaraokeStartIn": {
        "properties": {
          "suggestionId": {
            "type": "integer",
            "title": "Suggestionid"
          }
        },
        "type": "object",
        "required": [
          "suggestionId"
        ],
        "title": "KaraokeStartIn",
        "description": "🔴 O `suggestionId` no corpo não é decorativo: sem ele, um toque atrasado no botão do turno\nanterior iniciaria a vez de outra pessoa. Mesma função do `playId` no voto de skip."
      },
      "KaraokeStartOut": {
        "properties": {
          "playId": {
            "type": "integer",
            "title": "Playid"
          }
        },
        "type": "object",
        "required": [
          "playId"
        ],
        "title": "KaraokeStartOut"
      },
...
      "SearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/SearchResult"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "SearchResponse"
      },
      "SearchResult": {
        "properties": {
          "trackId": {
            "type": "string",
            "title": "Trackid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "artists": {
            "type": "string",
            "title": "Artists"
          },
          "album": {
            "type": "string",
            "title": "Album"
          },
          "artUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arturl"
          },
          "durationMs": {
            "type": "integer",
            "title": "Durationms"
          },
          "provider": {
            "type": "string",
            "enum": [
              "spotify",
              "karaoke"
            ],
            "title": "Provider",
            "default": "spotify"
          },
          "explicit": {
            "type": "boolean",
            "title": "Explicit"
          },
          "queueable": {
            "type": "boolean",
            "title": "Queueable",
            "default": true
          },
          "blockedReason": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "ALREADY_QUEUED",
                  "PLAYED_RECENTLY",
                  "TOO_LONG"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockedreason"
          },
          "blockedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blockedby"
          }
        },
        "type": "object",
        "required": [
          "trackId",
          "name",
          "artists",
          "album",
          "artUrl",
          "durationMs",
          "explicit"
        ],
        "title": "SearchResult"
      },
...
      "SuggestIn": {
        "properties": {
          "trackId": {
            "type": "string",
            "title": "Trackid"
          }
        },
        "type": "object",
        "required": [
          "trackId"
        ],
        "title": "SuggestIn"
      },
      "SuggestOut": {
        "properties": {
          "suggestionId": {
            "type": "integer",
            "title": "Suggestionid"
          },
          "positionHint": {
            "type": "string",
            "title": "Positionhint"
          },
          "cooldownUntilMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cooldownuntilms"
          }
        },
        "type": "object",
        "required": [
          "suggestionId",
          "positionHint",
          "cooldownUntilMs"
        ],
        "title": "SuggestOut"
      },
...
      "TvClaimIn": {
        "properties": {
          "tvId": {
            "type": "string",
            "maxLength": 64,
            "minLength": 8,
            "title": "Tvid"
          }
        },
        "type": "object",
        "required": [
          "tvId"
        ],
        "title": "TvClaimIn",
        "description": "O `tvId` é gerado pelo CLIENTE e vive no `sessionStorage` da aba.\n\n🔴 Gerado no cliente e não no servidor de propósito: `sessionStorage` sobrevive ao F5 da mesma\naba, então recarregar a `/tv` no meio de um karaokê reapresenta o mesmo id e a posse volta na\nhora. Com um id emitido pelo servidor a cada `claim`, uma recarga seria indistinguível de uma\nsegunda tela e o monitor perderia o som para si mesmo por 25 s."
      },
      "TvClaimOut": {
        "properties": {
          "owner": {
            "type": "boolean",
            "title": "Owner"
          }
        },
        "type": "object",
        "required": [
          "owner"
        ],
        "title": "TvClaimOut"
      },
      "TvReportIn": {
        "properties": {
          "playId": {
            "type": "integer",
            "title": "Playid"
          },
          "tvId": {
            "type": "string",
            "maxLength": 64,
            "title": "Tvid"
          },
          "state": {
            "type": "string",
            "enum": [
              "playing",
              "paused",
              "ended",
              "error"
            ],
            "title": "State"
          },
          "positionMs": {
            "type": "integer",
            "maximum": 86400000.0,
            "minimum": 0.0,
            "title": "Positionms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "playId",
          "tvId",
          "state",
          "positionMs"
        ],
        "title": "TvReportIn",
        "description": "O que a /tv reporta. Fronteira NÃO confiável e sem cookie (a /tv não tem e não vai ter):\nvalidação estrita é o controle compensatório, e é barata."
      },
      "TvReportOut": {
        "properties": {
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          }
        },
        "type": "object",
        "required": [
          "accepted"
        ],
        "title": "TvReportOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}